+++ src/api/interfaces/CRUDApi.ts
@@ -1,6 +1,8 @@
 export default interface CRUDApi {
 
-    fetchAll(orderBy: String, orderWay: String, page: Number, perPage: Number): Promise<String>;
+    fetchItems(orderBy: String, orderWay: String, page: Number, perPage: Number): Promise<String>;
+
+    fetchItem(id: String): Promise<Object>;
 
     post(data: Object): Promise<Object|String>;