+++ src/js/components/common/Datatable.js
@@ -7,13 +7,13 @@ export const Datatable = (props) => {
     const [orderBy, setOrderBy] = useState('id');
     const [orderWay, setOrderWay] = useState('asc');
     const [page, setPage] = useState(1);
-    const [perPage, setPerPage] = useState(1);
+    const [perPage, setPerPage] = useState(10);
 
-    const { fetch } = props;
+    const { fetchItems } = props;
 
     useEffect(() => {
-        fetch(orderBy, orderWay, page, perPage);
-    }, [fetch, orderBy, orderWay, page, perPage])
+        fetchItems(orderBy, orderWay, page, perPage);
+    }, [fetchItems, orderBy, orderWay, page, perPage])
 
     const handleSort = (column, sortDirection) => {
         const nextOrderBy = column.selector;
@@ -34,6 +34,14 @@ export const Datatable = (props) => {
         setPage(nextPage);
     }
 
+    const handleRowClicked = ({id}, e) => {
+        e.stopPropagation();
+        if (e.target.type === 'button') {
+            props.openModalDetail(id);
+        }
+        return;
+    }
+
     return (
 
         <DataTable
@@ -45,7 +53,8 @@ export const Datatable = (props) => {
             paginationServer
             paginationTotalRows={props.total}
             paginationPerPage={perPage}
-            paginationRowsPerPageOptions={[1, 10, 15, 20, 25, 30]}
+            paginationRowsPerPageOptions={[10, 15, 20, 25, 30]}
+            onRowClicked={handleRowClicked}
             onSort={handleSort}
             onChangeRowsPerPage={handleChangeRowsPerPage}
             onChangePage={handleChangePage}