+++ src/js/components/presentationals/RolesPresentational.js
@@ -1,27 +1,29 @@
 import { Fragment, useEffect, useState } from 'react';
-import { Card } from 'react-bootstrap';
+import { Button, Card } from 'react-bootstrap';
 import { connect } from "react-redux";
 import { actions as refreshActions } from '../../actions/RefreshActions';
 import { actions as rolesActions } from "../../actions/RolesAction";
 import Datatable from "../common/Datatable"
+import ModalWindow from '../common/ModalWindow';
+import RoleDetail from '../details/RoleDetail';
 
 export const RolesPresentational = props => {
 
-  const [mounted, setMounted] = useState(false);
+    const [mounted, setMounted] = useState(false);
 
-  useEffect(() => {
-    if (!mounted) {
-        props.refreshToken();
-        setMounted(true);
-    }
-  }, [mounted, props])
-
+    useEffect(() => {
+        if (!mounted) {
+            props.refreshToken();
+            setMounted(true);
+        }
+    }, [mounted, props])
 
     const columns = [
         {
             name: 'ID',
             selector: 'id',
             sortable: true,
+            cell: row => <Button variant="link" data-tag="allowRowEvents">{row.id}</Button>
         },
         {
             name: 'Role',
@@ -47,29 +49,44 @@ export const RolesPresentational = props => {
                 <Card.Body>
                     <Datatable
                         columns={columns}
-                        data={props.roles}
+                        data={props.items}
                         {...props}
                     />
                 </Card.Body>
             </Card>
+            <ModalWindow
+                show={props.show}
+                componentDetail={
+                    <RoleDetail
+                        item={props.item}
+                    />
+                }
+                {...props}
+            />
         </Fragment>
   );
 };
 
 const mapStateToProps = state => {
     return {
-      isLoading: state.roles && state.roles.isLoading ? state.roles.isLoading : null,
-      roles: state.roles.roles,
-      total: state.roles.total,
+        isLoading: state.roles && state.roles.isLoading ? state.roles.isLoading : null,
+        items: state.roles.items,
+        item: state.roles.item,
+        total: state.roles.total,
+        show: state.roles.showModal,
+        idDetail: state.roles.idDetail,
     };
   }
   
-  const mapDispatchToProps = dispatch => {
+const mapDispatchToProps = dispatch => {
     return {
-      fetch: (orderBy, orderWay, page, perPage) => dispatch(rolesActions.fetchAll(orderBy, orderWay, page, perPage)),
-      refreshToken: () => dispatch(refreshActions.refreshToken()),
+        fetchItems: (orderBy, orderWay, page, perPage) => dispatch(rolesActions.fetchItems(orderBy, orderWay, page, perPage)),
+        fetchItem: (id) => dispatch(rolesActions.fetchItem(id)),
+        openModalDetail: (id) => dispatch(rolesActions.openModalDetail(id)),
+        closeModalDetail: () => dispatch(rolesActions.closeModalDetail()),
+        refreshToken: () => dispatch(refreshActions.refreshToken()),
     }
-  }
+}
   
 export default connect(
     mapStateToProps,